#4706: refactoring
