+++ src/app/ubs/ubs-admin/models/not-taken-out-reason.model.ts
@@ -1,0 +1,10 @@
+export interface NotTakenOutReasonImage {
+  src: string;
+  name: string | null;
+  file: File;
+}
+
+export interface DataToSend {
+  description: string;
+  images?: string[] | null;
+}
